Rename to vpn service mod
