+++ nym-vpnd/src/service/mod.rs
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 mod exit_listener;
-mod service;
+mod vpn_service;
 mod start;
 mod status_listener;
 
-pub(crate) use service::{
+pub(crate) use vpn_service::{
     VpnServiceCommand, VpnServiceConnectResult, VpnServiceDisconnectResult, VpnServiceStatusResult,
 };
 pub(crate) use start::start_vpn_service;
